Moved send and recv into send_recv.py
